requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
openai==1.3.0
httpx==0.24.1
python-dotenv==1.0.0
flask==3.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
cryptography==41.0.7
werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
numpy<2.0
pandas>=2.0.0,<2.1.0
openpyxl==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
boto3==1.34.0
celery==5.3.4
redis==5.0.1
kombu==5.3.4
